refactor: WaitingFrameCount -> WaitingPitchingCount로 리네이밍
